pyproject.toml
[project]
name = "grpc-image-gen"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "flatbuffers>=25.12.19",
    "fpzip>=1.2.5",
    "grpcio>=1.80.0",
    "grpcio-tools>=1.80.0",
    "numpy>=2.4.4",
    "pillow>=12.2.0",
    "pyyaml>=6.0.3",
]

[project.scripts]
drawthings-grpc-generate = "drawthings_grpc_sample.generate:main"
drawthings-grpc-model-list = "drawthings_grpc_sample.model_list:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/drawthings_grpc_sample"]
